adding userInit class -- global state
