+++ src/Register.js
@@ -17,115 +17,15 @@ import './css/open-sans.css'
 import './css/pure-min.css'
 import './App.css'
 
-class Register_1 extends Component {
+import userInitialisation from './userInitialisation';
+
+class userRegistration extends userInitialisation {
     constructor(props) {
         super(props) ;
 
-        this.state = {
-            storageValue: 45,
-            account: 0,
-            web3: null,
-            registered: false,
-            name: 'n/a',
-            surname: 'n/a',
-            email: 'n/a',
-            owner: 'n/a',
-            formName: '',
-            formSurname: '',
-            formEmail: ''
-        };
-
-        this.userDetail = {
-            name: null,
-            surname: null,
-            email: null,
-            omneeIDAddress: null
-        };
 
         this.handleChange = this.handleChange.bind(this);
         this.handleSubmit = this.handleSubmit.bind(this);
-    }
-
-    componentWillMount() {
-        // Get network provider and web3 instance.
-        // See utils/getWeb3 for more info.
-        getWeb3
-            .then(results => {
-                this.setState({
-                    web3: results.web3,
-                    account: results.web3.eth.accounts[0]
-                })
-
-                // Instantiate contract once web3 provided.
-                this.instantiateContract()
-            })
-            .catch(() => {
-                console.log('Error finding web3.')
-            })
-    }
-
-    instantiateContract() {
-        /*
-         * SMART CONTRACT EXAMPLE
-         *
-         * Normally these functions would be called in the context of a
-         * state management library, but for convenience I've placed them here.
-         */
-
-        const contract = require('truffle-contract');
-
-        const omneePortal = contract(omneePortalContract);
-        const omneeID = contract(omneeIDContract);
-
-        omneePortal.setProvider(this.state.web3.currentProvider);
-        omneeID.setProvider(this.state.web3.currentProvider);
-
-        // Declaring this for later so we can chain functions
-        var omneePortalInstance, omneeIDInstance;
-
-        // Record the current account holder to display on page
-        this.setState({account: this.state.web3.eth.accounts[0]});
-
-        // Get accounts.
-        this.state.web3.eth.getAccounts((error, accounts) => {
-            omneePortal.deployed().then((instance) => {
-                omneePortalInstance = instance;
-                //return omneePortalInstance.set(5, {from: accounts[0]})
-                return omneePortalInstance.userExists.call({from: accounts[0]})
-            }).then((result) => {
-                if(result.c[0] === 1) // user doesn't exist, we need to create an id
-                {
-                    console.log('User doesnt exist, attempting to create user');
-                    this.setState({registered: false});
-                } else {
-                    console.log('User exists, moving on...');
-                    this.setState({registered: true});
-                }
-            }).then((result) => {
-                console.log('Getting omneeID address');
-                // get the omneeID ID
-                return omneePortalInstance.id.call({from: accounts[0]})
-            }).then((result) => {
-                console.log(result)
-                this.userDetail.omneeIDAddress = result;
-                console.log('Connecting to omneeID contract to get info');
-                omneeIDInstance = omneeID.at(result);
-                return omneeIDInstance.getInfo.call({from: accounts[0]})
-            }).then((result) => {
-                console.log(result)
-                if (result === undefined || result === null || result[0] === "") {
-                    console.log('No user info found')
-                } else {
-                    console.log('User info found');
-                    this.setState({
-                        name: result[0],
-                        surname: result[1],
-                        email: result[2],
-                        owner: result[3]
-                    })
-                }
-            })
-        })
     }
 
     handleSubmit(event) {
@@ -160,30 +60,34 @@ class Register_1 extends Component {
 
     registerForm() {
         return (
-            <form className="pure-form pure-form-aligned" onSubmit={this.handleSubmit}>
-                <div className="pure-control-group">
-                    <label>First Name</label>
-                    <input id="firstname" type="text"
-                           placeholder="First Name" value={this.state.formName} onChange={this.handleChange} />
-                    <span className="pure-form-message-inline">This is a required field.</span>
+          <div className="row justify-content-md-center">
+            <div className="registerForm col-md-5">
+              <h4>Welcome, please enter your details</h4>
+              Your Ethereum address: <br />
+              <p className="address">{this.userDetails.omneeIDAddress}</p>
+              <form onSubmit={this.handleSubmit}>
+                <div className="form-group">
+                  <label>First name</label>
+                  <input type="email" className="form-control" id="registerFirstName" placeholder="e.g. Alice" value="" onChange={this.handleChange} />
                 </div>
-
-                <div className="pure-control-group">
-                    <label>Second Name</label>
-                    <input id="secondname" type="text" placeholder="Second Name" value={this.state.formSurname}/>
+                <div className="form-group">
+                  <label>Surname</label>      
+                  <input type="text" className="form-control" id="registerSurname" placeholder="e.g. Smith" value="" />
                 </div>
-
-                <div className="pure-control-group">
-                    <label>Email Address</label>
-                    <input id="email" type="email" placeholder="Email Address"/>
+                <div className="form-group">
+                  <label>Email Address</label>
+                  <input id="email" className="form-control" type="email" placeholder="e.g. a.smith@gmail.com"/>
                 </div>
-
-                <div className="pure-controls">
-                    <button type="submit" className="pure-button pure-button-primary">Submit</button>
+                <div className="form-group">
+                  <label>NHS Number</label>
+                  <input id="registerNHSNumber" className="form-control" placeholder="e.g. 987 654 4321" type="text"/>
                 </div>
-            </form>
+                <button type="submit" className="btn btn-primary">Submit</button>
+              </form>
+            </div>
+          </div>
         )
-    }
+      }
 
     render() {
         const form = this.registerForm() ;
@@ -193,10 +97,10 @@ class Register_1 extends Component {
                     <div className="column middle">
                                     <h3>Your details are:</h3>
                                     <ul>
-                                        <li><strong>First name:</strong> {this.state.name}</li>
-                                        <li><strong>Surname and other name(s):</strong> {this.state.surname}</li>
-                                        <li><strong>Email address:</strong> {this.state.email}</li>
-                                        <li><strong>Owner add:</strong> {this.state.owner}</li>
+                                        <li><strong>First name:</strong> {this.userDetails.name}</li>
+                                        <li><strong>Surname and other name(s):</strong> {this.userDetails.surname}</li>
+                                        <li><strong>Email address:</strong> {this.userDetails.email}</li>
+                                        <li><strong>Owner add:</strong>{this.userDetails.omneeIDAddress}</li>
                                     </ul>
                                     {form}
 
@@ -207,4 +111,4 @@ class Register_1 extends Component {
     }
 }
 
-export default Register_1 ;
+export default userRegistration ;
